pyproject.toml
[project]
name = "open-rag-layer"
version = "0.1.0"
description = "Provider-agnostic multimodal retrieval layer with Gemini Embedding 2 and Qdrant"
requires-python = ">=3.12"

dependencies = [
  "google-genai>=1.0",
  "qdrant-client>=1.9",
  "pydantic>=2.0",
  "pypdf>=4.0",
  "pillow>=10.0",
  "numpy>=1.26",
  "rank-bm25>=0.2",
  "loguru>=0.7",
  "httpx>=0.27",
  "filetype>=1.2",
  "soundfile>=0.12",
]

[dependency-groups]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "pytest-mock>=3.12"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rag_layer"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
